+++ flow-core/src/main/java/com/dragon/flow/service/org/impl/DepartmentServiceImpl.java
@@ -131,7 +131,7 @@ public class DepartmentServiceImpl extends ServiceImpl<IDepartmentMapper, Depart
         ids.add(id);
         LambdaQueryWrapper<Department> departmentLambdaQueryWrapper = new LambdaQueryWrapper<>();
         departmentLambdaQueryWrapper.eq(Department::getPid, id).eq(Department::getDelFlag, FlowConstant.DEL_FLAG_1);
-        int count = this.count(departmentLambdaQueryWrapper);
+        long count = this.count(departmentLambdaQueryWrapper);
         if (count > 0){
             List<Department> companyList = this.list(departmentLambdaQueryWrapper);
             List<String> finalIds = ids;
@@ -159,14 +159,14 @@ public class DepartmentServiceImpl extends ServiceImpl<IDepartmentMapper, Depart
         ReturnVo<String> returnVo = new ReturnVo<>(ReturnCode.SUCCESS, "OK");
         LambdaQueryWrapper<Personal> personalLambdaQueryWrapper = new LambdaQueryWrapper<>();
         personalLambdaQueryWrapper.in(Personal::getDeptId, ids).eq(Personal::getDelFlag, FlowConstant.DEL_FLAG_1);
-        int count = personalService.count(personalLambdaQueryWrapper);
+        long count = personalService.count(personalLambdaQueryWrapper);
         if (count > 0){
             returnVo = new ReturnVo<>(ReturnCode.FAIL, "请先移除部门相关人员!");
         } else {
             LambdaQueryWrapper<Department> departmentLambdaQueryWrapper = new LambdaQueryWrapper<>();
             departmentLambdaQueryWrapper.eq(Department::getDelFlag, FlowConstant.DEL_FLAG_1)
                     .eq(Department::getPid, ids.get(0));
-            int departmentCount = this.count(departmentLambdaQueryWrapper);
+            long departmentCount = this.count(departmentLambdaQueryWrapper);
             if (departmentCount > 0){
                 returnVo = new ReturnVo<>(ReturnCode.FAIL, "该部门还存在子部门，请确认!");
             } else {
